f/mmcdaniel-baroreceptors: Fix anesthesia machine activation
When the anesthesia machine was activated, the tidal volume would
decrease to near 0 even when a mask was applied because
outflow from the lung was stopped during the first few breathing cycles
with the machine active.  This occured because the combined
respiratory/AM circuit is never run to a steady state.  Thus, the first few patient breaths with the machine active were being used to charge the ventilator compliance element and were not going to the lungs.  Running out the AM circuit for a few cycles during initialization (similar to TuneCircuit in Respiratory and Cardiovascular) resolved this issue.